Refactored execution to a separate module
